refactor: rearrange some code
