+++ src/swarm-lang/Swarm/Language/Typecheck.hs
@@ -66,14 +66,13 @@ import Data.Maybe
 import Data.Set (Set, (\\))
 import Data.Set qualified as S
 import Data.Text qualified as T
-import Swarm.Effect.Unify (Unification, (=:=))
+import Swarm.Effect.Unify (Unification, (=:=), UnificationError)
 import Swarm.Effect.Unify qualified as U
 import Swarm.Language.Context hiding (lookup)
 import Swarm.Language.Context qualified as Ctx
 import Swarm.Language.Module
 import Swarm.Language.Parse.QQ (tyQ)
 import Swarm.Language.Syntax
-import Swarm.Language.Typecheck.Unify (UnificationError)
 import Swarm.Language.Typecheck.Unify qualified as U
 import Swarm.Language.Types
 import Prelude hiding (lookup)
