+++ swarm.cabal
@@ -137,7 +137,6 @@ library swarm-lang
     Swarm.Language.Text.Markdown
     Swarm.Language.Typecheck
     Swarm.Language.Typecheck.Unify
-    Swarm.Language.Typecheck.Unify.Subst
     Swarm.Language.Typed
     Swarm.Language.Types
     Swarm.Language.Value
@@ -637,7 +636,6 @@ library
     Swarm.Language.Text.Markdown,
     Swarm.Language.Typecheck,
     Swarm.Language.Typecheck.Unify,
-    Swarm.Language.Typecheck.Unify.Subst,
     Swarm.Language.Typed,
     Swarm.Language.Types,
     Swarm.Language.Value,
